requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
discord==2.1.0
discord.py==2.1.0
executing==1.2.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
multidict==6.0.4
numpy==1.24.1
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.36
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
SQLAlchemy==1.4.46
stack-data==0.6.2
traitlets==5.8.1
wcwidth==0.2.6
Werkzeug==2.2.2
yarl==1.8.2
